# api/swagger_webservice.yaml
openapi: 3.1.0
info:
  title: Dimension Game API
  version: 1.1.0
  description: The Dimension Game API allows users to create and manage games, initiate rounds, and monitor game progress including player statuses and scores.
  license:
    name: Copyright
servers:
  - url: https://localhost:8080/
...
paths:
...
  /training/{trainID}:
    patch:
      tags:
        - Training
      summary: Play the training session
      operationId: playTrainingSession
      parameters:
        - name: trainID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Submit a Dimension for the training session
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dimension'
      responses:
        202:
          description: accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingSession'

    get:
      tags:
        - Training
      summary: Retrieve training session status
      operationId: getTrainingStatus
      parameters:
        - name: trainID
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Training session status
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: number
                  bonusPoints:
                    type: boolean
                  submittedDimension:
                    $ref: '#/components/schemas/Dimension'
                  tasks:
                    type: array
                    items:
                      type: string
                  taskViolations:
                    type: array
                    items:
                      type: string
                  expirationTime:
                    type: string
                    format: date-time
  /training/{trainID}/regenerate:
    post:
      tags:
        - Training
      summary: regenerate the training session
      operationId: regenerateTrainingSession
      parameters:
        - name: trainID
          in: path
          required: true
          schema:
            type: string
      responses:
        202:
          description: accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      type: string
components:
  schemas:
...
    Dimension:
      type: object
      properties:
        a: { type: string, enum: [G, K, O, W, B] }
        b: { type: string, enum: [G, K, O, W, B] }
        c: { type: string, enum: [G, K, O, W, B] }
        d: { type: string, enum: [G, K, O, W, B] }
        e: { type: string, enum: [G, K, O, W, B] }
        f: { type: string, enum: [G, K, O, W, B] }
        g: { type: string, enum: [G, K, O, W, B] }
        h: { type: string, enum: [G, K, O, W, B] }
        i: { type: string, enum: [G, K, O, W, B] }
        j: { type: string, enum: [G, K, O, W, B] }
        k: { type: string, enum: [G, K, O, W, B] }
        l: { type: string, enum: [G, K, O, W, B] }
        m: { type: string, enum: [G, K, O, W, B] }
        n: { type: string, enum: [G, K, O, W, B] }
...
    TrainingSession:
      type: object
      properties:
        score:
          type: number
        bonusPoints:
          type: boolean
        submittedDimension:
          $ref: '#/components/schemas/Dimension'
        tasks:
          type: array
          items:
            type: string
        taskViolations:
          type: array
          items:
            type: string
        expirationTime:
          type: string
          format: date-time